edited code

+++ 02.DataTypesAndVariables-Lab/03.ExactSumOfRealNumbers/Program.cs
@@ -3,11 +3,11 @@
     private static void Main(string[] args)
     {
         int n = int.Parse(Console.ReadLine());
-        double sum = 0;
+        decimal sum = 0;
 
         for (int i = 1; i <= n; i++)
         {
-            double currentNumber = double.Parse(Console.ReadLine());
+            decimal currentNumber = decimal.Parse(Console.ReadLine());
             sum += currentNumber;
         }
